Simplify and encapsulate window function state management
